# artifacts/zenflow/backend/internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Action defines model for Action.
type Action struct {
	ActionType      *string                 `json:"action_type,omitempty"`
	Id              *int64                  `json:"id,omitempty"`
	PayloadTemplate *map[string]interface{} `json:"payload_template,omitempty"`
	SequenceOrder   *int                    `json:"sequence_order,omitempty"`
	WorkflowId      *int64                  `json:"workflow_id,omitempty"`
}

// Execution defines model for Execution.
type Execution struct {
	CreditsSpent *int       `json:"credits_spent,omitempty"`
	ExecutedAt   *time.Time `json:"executed_at,omitempty"`
	Id           *int64     `json:"id,omitempty"`
	OrgId        *int64     `json:"org_id,omitempty"`
	Status       *string    `json:"status,omitempty"`
	WorkflowId   *int64     `json:"workflow_id,omitempty"`
}

// Organization defines model for Organization.
type Organization struct {
	CreditsBalance *int    `json:"credits_balance,omitempty"`
	Id             *int64  `json:"id,omitempty"`
	Name           *string `json:"name,omitempty"`
	PlanType       *string `json:"plan_type,omitempty"`
}

// User defines model for User.
type User struct {
	Email *string `json:"email,omitempty"`
	Id    *int64  `json:"id,omitempty"`
	OrgId *int64  `json:"org_id,omitempty"`
	Role  *string `json:"role,omitempty"`
}

// Workflow defines model for Workflow.
type Workflow struct {
	CreatedAt    *time.Time `json:"created_at,omitempty"`
	Id           *int64     `json:"id,omitempty"`
	OrgId        *int64     `json:"org_id,omitempty"`
	Status       *string    `json:"status,omitempty"`
	Title        *string    `json:"title,omitempty"`
	TriggerEvent *string    `json:"trigger_event,omitempty"`
}

// LoginJSONBody defines parameters for Login.
type LoginJSONBody struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// RegisterJSONBody defines parameters for Register.
type RegisterJSONBody struct {
	Email    string `json:"email"`
	OrgId    int64  `json:"org_id"`
	Password string `json:"password"`
	Role     string `json:"role"`
}

// CreateOrganizationJSONBody defines parameters for CreateOrganization.
type CreateOrganizationJSONBody struct {
	CreditsBalance int    `json:"credits_balance"`
	Name           string `json:"name"`
	PlanType       string `json:"plan_type"`
}

// CreateWorkflowJSONBody defines parameters for CreateWorkflow.
type CreateWorkflowJSONBody struct {
	Title        string `json:"title"`
	TriggerEvent string `json:"trigger_event"`
}

// CreateActionJSONBody defines parameters for CreateAction.
type CreateActionJSONBody struct {
	ActionType      string                  `json:"action_type"`
	PayloadTemplate *map[string]interface{} `json:"payload_template,omitempty"`
	SequenceOrder   int                     `json:"sequence_order"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody LoginJSONBody

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody RegisterJSONBody

// CreateOrganizationJSONRequestBody defines body for CreateOrganization for application/json ContentType.
type CreateOrganizationJSONRequestBody CreateOrganizationJSONBody

// CreateWorkflowJSONRequestBody defines body for CreateWorkflow for application/json ContentType.
type CreateWorkflowJSONRequestBody CreateWorkflowJSONBody

// CreateActionJSONRequestBody defines body for CreateAction for application/json ContentType.
type CreateActionJSONRequestBody CreateActionJSONBody