# specs/zenflow/api/openapi.yaml
openapi: 3.0.3
info:
  title: ZenFlow API
  version: 1.0.0
  description: Multi-tenant workflow automation SaaS

servers:
  - url: http://localhost:8080

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Organization:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        plan_type:
          type: string
        credits_balance:
          type: integer

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        org_id:
          type: integer
          format: int64
        email:
          type: string
        role:
          type: string

    Workflow:
      type: object
      properties:
        id:
          type: integer
          format: int64
        org_id:
          type: integer
          format: int64
        title:
          type: string
        trigger_event:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time

    Action:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workflow_id:
          type: integer
          format: int64
        action_type:
          type: string
        payload_template:
          type: object
        sequence_order:
          type: integer

    Execution:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workflow_id:
          type: integer
          format: int64
        org_id:
          type: integer
          format: int64
        status:
          type: string
        credits_spent:
          type: integer
        executed_at:
          type: string
          format: date-time

paths:
  /auth/register:
    post:
      operationId: Register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_id, email, password, role]
              properties:
                org_id:
                  type: integer
                  format: int64
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'

  /auth/login:
    post:
      operationId: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Invalid credentials

  /organizations:
    post:
      operationId: CreateOrganization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, plan_type, credits_balance]
              properties:
                name:
                  type: string
                plan_type:
                  type: string
                credits_balance:
                  type: integer
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  organization:
                    $ref: '#/components/schemas/Organization'

  /workflows:
    post:
      operationId: CreateWorkflow
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, trigger_event]
              properties:
                title:
                  type: string
                trigger_event:
                  type: string
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow:
                    $ref: '#/components/schemas/Workflow'
        '403':
          description: Permission denied
        '404':
          description: User not found

    get:
      operationId: ListWorkflows
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workflows listed
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'
        '403':
          description: Permission denied
        '404':
          description: User not found

  /workflows/{id}:
    get:
      operationId: GetWorkflow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow:
                    $ref: '#/components/schemas/Workflow'
        '403':
          description: Permission denied
        '404':
          description: Workflow not found

  /workflows/{id}/activate:
    post:
      operationId: ActivateWorkflow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow:
                    $ref: '#/components/schemas/Workflow'
        '402':
          description: Insufficient credits
        '403':
          description: Permission denied
        '404':
          description: Workflow not found
        '409':
          description: Invalid state transition

  /workflows/{id}/pause:
    post:
      operationId: PauseWorkflow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow paused
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow:
                    $ref: '#/components/schemas/Workflow'
        '403':
          description: Permission denied
        '404':
          description: Workflow not found
        '409':
          description: Invalid state transition

  /workflows/{id}/archive:
    post:
      operationId: ArchiveWorkflow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow:
                    $ref: '#/components/schemas/Workflow'
        '403':
          description: Permission denied
        '404':
          description: Workflow not found
        '409':
          description: Invalid state transition

  /workflows/{id}/execute:
    post:
      operationId: ExecuteWorkflow
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution:
                    $ref: '#/components/schemas/Execution'
        '403':
          description: Permission denied
        '404':
          description: Workflow not found
        '409':
          description: Workflow not active

  /workflows/{id}/actions:
    post:
      operationId: CreateAction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action_type, sequence_order]
              properties:
                action_type:
                  type: string
                payload_template:
                  type: object
                sequence_order:
                  type: integer
      responses:
        '201':
          description: Action created
          content:
            application/json:
              schema:
                type: object
                properties:
                  action:
                    $ref: '#/components/schemas/Action'
        '403':
          description: Permission denied
        '404':
          description: Workflow not found

    get:
      operationId: ListActions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Actions listed
          content:
            application/json:
              schema:
                type: object
                properties:
                  actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Action'
        '403':
          description: Permission denied
        '404':
          description: Workflow not found